ai_learn/text2sql/requirements.txt
# Text2SQL 演示项目依赖
# 更新到支持 Python 3.13 的版本

# 核心框架
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0

# 数据库
sqlalchemy>=2.0.0

# Web界面
streamlit>=1.31.0

# 环境变量管理
python-dotenv>=1.0.0

# 数据处理
pandas>=2.0.0
